+++ resources/js/views/client_portal/ViewInvoice.vue
@@ -19,6 +19,15 @@
                         icon="fab fa-paypal"
                         @click="payInvoiceWithGateway('paypal')"
                     />
+
+                    <Button
+                        v-if="$settings.coinbase_available && invoice.status != 'paid' && invoice.status != 'overpaid'"
+                        id="coinbase_button"
+                        v-tooltip:top="$t('invoice.click_for_pay_with_coinbase')"
+                        class="mr-2 no-print"
+                        icon="fab fa-bitcoin"
+                        @click="payInvoiceWithGateway('coinbase')"
+                    />
                     <Button
                         v-if="invoice.download"
                         :label="$t('basic.download')"